refactor: address PR review feedback
- Fix LRCLib duration regression: protocol now takes Int? (was Double?),
  matching the previous AF behavior where Double duration was coerced to
  whole seconds before being sent. The DataSource layer truncates via
  Int.init at the call site. (Codex P2)
- Test quality: convert URL-construction tests in LyricsDataSourceTests
  from try? to async throws + try await so request-construction failures
  surface. Search-side tests use a `[]` body to satisfy decode. Add a
  regression test verifying duration is sent as integer seconds.
  (CodeRabbit)
- DRY: derive HealthCheck probe URLs from each protocol's baseURL constant
  and shared userAgent constant, instead of hard-coded literals. Keeps the
  health check coupled to the API declaration. (CodeRabbit nit)
- OpenAICompatibleHealthCheck: surface JSON serialization failures
  explicitly instead of silently sending an empty body. (CodeRabbit nit)
- OpenAICompatibleAPITests: simplify the trailing-slash test by directly
  inspecting provider(for:).baseURL instead of building two providers
  with a recorder. (CodeRabbit nit)

+++ Tests/LyricsDataSourceTests/Helpers/LRCLibStub.swift
@@ -6,11 +6,11 @@ import Foundation
 /// Manual mock of the `LRCLib` protocol for testing `LyricsDataSourceImpl`
 /// without exercising URL construction or networking.
 struct LRCLibStub: LRCLib, @unchecked Sendable {
-    let getResult: @Sendable (_ trackName: String, _ artistName: String, _ duration: Double?) async throws -> LyricsResult
+    let getResult: @Sendable (_ trackName: String, _ artistName: String, _ duration: Int?) async throws -> LyricsResult
     let searchResult: @Sendable (_ q: String) async throws -> [LyricsResult]
 
     init(
-        get: @escaping @Sendable (_ trackName: String, _ artistName: String, _ duration: Double?) async throws -> LyricsResult = { _, _, _ in .empty
+        get: @escaping @Sendable (_ trackName: String, _ artistName: String, _ duration: Int?) async throws -> LyricsResult = { _, _, _ in .empty
         },
         search: @escaping @Sendable (_ q: String) async throws -> [LyricsResult] = { _ in [] }
     ) {
@@ -18,7 +18,7 @@ struct LRCLibStub: LRCLib, @unchecked Sendable {
         self.searchResult = search
     }
 
-    func get(trackName: String, artistName: String, duration: Double?) async throws -> LyricsResult {
+    func get(trackName: String, artistName: String, duration: Int?) async throws -> LyricsResult {
         try await getResult(trackName, artistName, duration)
     }
